requirements.txt
apache-airflow-providers-snowflake==5.4.0
snowflake-connector-python==3.6.0
dbt-snowflake==1.7.0